Move libpromises/crypto.[ch] to libcfcrypto
Except for the few things that have to stay in libpromises which
where moved to generic_agent.[ch] (one function) and
lastseen_crypto.[ch].
