+++ cf-serverd/server_common.c
@@ -31,6 +31,7 @@ static const int CF_NOSIZE = -1;
 #include <item_lib.h>                                 /* ItemList2CSV_bound */
 #include <string_lib.h>                              /* ToLower,StrCatDelim */
 #include <regex.h>                                    /* StringMatchFull */
+#include <crypto_init.h>                              /* CF_DEFAULT_DIGEST, CfengineCipher */
 #include <crypto.h>                                   /* EncryptString */
 #include <files_names.h>
 #include <files_interfaces.h>
