+++ libpromises/cf3.extern.h
@@ -58,8 +58,6 @@ extern int VEXPIREAFTER;
 extern const char *const OBSERVABLES[CF_OBSERVABLES][2];
 
 extern bool FIPS_MODE;
-extern HashMethod CF_DEFAULT_DIGEST;
-extern int CF_DEFAULT_DIGEST_LEN;
 
 extern int CF_PERSISTENCE;
